+++ school/schema.py
@@ -11,7 +11,7 @@ from school.apps.authentication import schema as auth_schema
 class Mutation(
     auth_schema.Mutation, graphene.ObjectType
 ):
-    token_auth = graphql_jwt.ObtainJSONWebToken().Field()
+    # token_auth = graphql_jwt.ObtainJSONWebToken().Field()
     verify_token = graphql_jwt.Verify.Field()
     refresh_token = graphql_jwt.Refresh.Field()
